ExceptionHandler->target should be int32_t, not intptr_t (bug 500181, r=lhansen+)
The reason it used to be intptr_t was to handle MIR, which overwrite
the (int32) pc_offset with an (intptr_t) machine code address.

LIR and the interpreters don't do this, and making it the same
type as ExceptionHandler.from/to cleans up code and shrinks ExceptionHandler.

+++ core/Verifier.cpp
@@ -2823,7 +2823,7 @@ namespace avmplus
 				{
 					core->console << "            exception["<<i<<"] from="<< handler->from
 						<< " to=" << handler->to
-						<< " target=" << (uint64_t)handler->target 
+						<< " target=" << handler->target 
 						<< " type=" << t
 						<< " name=";
 					if (name_index != 0)
@@ -2861,7 +2861,7 @@ namespace avmplus
 				WB(core->GetGC(), table, &handler->scopeTraits, scopeTraits);
 
 				
-				getFrameState((int)handler->target)->targetOfBackwardsBranch = true;
+				getFrameState(handler->target)->targetOfBackwardsBranch = true;
 
 				handler++;
 			}
